requirements.txt
# ctransformers==0.2.5
ctransformers
# sentence-transformers==2.2.2
sentence-transformers
pinecone-client
langchain
flask
langchain-community
PyPDF
langchain-chroma 
langchain-openai
langchain-pinecone
python-dotenv
-e .